fixed spec name and changed buildre spec
