--- apis/fluent-en_GB/atrium-api-fluent-en_GB-common/src/test/kotlin/ch/tutteli/atrium/api/fluent/en_GB/IterableContainsInOrderElementsOfAssertionSpec.kt
@@ -1,37 +1,0 @@
-package ch.tutteli.atrium.api.fluent.en_GB
-
-import ch.tutteli.atrium.creating.Expect
-import org.spekframework.spek2.Spek
-import kotlin.reflect.KFunction3
-
-class IterableContainsInOrderElementsOfAssertionSpec : Spek({
-    include(BuilderSpec)
-}) {
-    object BuilderSpec : ch.tutteli.atrium.specs.integration.IterableContainsInOrderOnlyEntriesAssertionsSpec  (
-        getContainsPair(),
-        getContainsNullablePair(),
-        "◆ ", "✔ ", "✘ ", "❗❗ ", "⚬ ", "» ", "▶ ", "◾ ",
-        "[Atrium][Builder] "
-    )
-
-    companion object : IterableContainsSpecBase() {
-        fun getContainsPair() = "$contains.$inOrder.$inOrderElementsOf" to Companion::containsInOrderOnlyValues
-
-        private fun containsInOrderOnlyValues(
-            expect: Expect<Iterable<Double>>,
-            a: Expect<Double>.() -> Unit,
-            aX: Array<out Expect<Double>.() -> Unit>
-        ): Expect<Iterable<Double>> = expect.contains.inOrder.only.elementsOf(listOf(a, *aX))
-
-        fun getContainsNullablePair() =
-            "$contains.$inOrder.$inOrderElementsOf" to Companion::containsInOrderOnlyNullableValues
-
-        private fun containsInOrderOnlyNullableValues(
-            expect: Expect<Iterable<Double?>>,
-            a: (Expect<Double>.() -> Unit)?,
-            aX: Array<out (Expect<Double>.() -> Unit)?>
-        ): Expect<Iterable<Double?>> = expect.contains.inOrder.only.elementsOf(listOf(a, *aX))
-
-    }
-}
-
